[API] Adjust adjustments contract tests after upmerge
